Added Freezing-Modifier  - the Frost-Walker-Modifier  - only for Boots (requested by DESM.al)
Fixed various Bugs with Aquaphilic.
Fixed Bug where Modifying would have deleted the item on the cursor (reported by Vicarious)

+++ src/main/java/de/flo56958/MineTinker/Listeners/AnvilListener.java
@@ -47,6 +47,7 @@ public class AnvilListener implements Listener {
         if (!(modManager.isToolViable(tool) || modManager.isArmorViable(tool))) { return; }
 
         boolean deleteAllItems = false;
+        if (e.getCursor() != null && !e.getCursor().getType().equals(Material.AIR)) { return; }
         if (!modManager.isModifierItem(modifier)) { //something else
             if (newTool != null && tool.getType().equals(newTool.getType())) { return; } //Not an upgrade
             deleteAllItems = true;
